reference engineError command => util
