--- core/command/engineErrors.js
@@ -1,13 +1,0 @@
-module.exports = {
-  execute: function (config) {
-    var compareConfig = require(config.tempCompareConfigFileName).compareConfig;
-    const error = compareConfig.testPairs.find(testPair => {
-      return !!testPair.engineErrorMsg;
-    });
-
-    if (error) {
-      return Promise.reject(error);
-    }
-    return Promise.resolve();
-  }
-};
