+++ examples/responsiveDemo/package.json
@@ -5,7 +5,8 @@
   "main": "index.html",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "node ./node_modules/super-simple-web-server/ ./"
+    "serve": "node ./node_modules/super-simple-web-server/ ./",
+    "reference": "node ../../cli/index.js reference"
   },
   "author": "",
   "license": "ISC",
